[IMP] *website_sale*: move variant_mixin.js into website_sale.js
task-5089578
