+++ addons/l10n_ar_website_sale/static/src/interactions/website_sale.js
@@ -3,9 +3,12 @@ import { WebsiteSale } from '@website_sale/interactions/website_sale';
 
 patch(WebsiteSale.prototype, {
     /**
-     * @override
-     * Updates the product's excluded price based on the selected variant.
-     * Ensuring availability info stays accurate.
+     * Override of `website_sale` to update the product's excluded price based on the selected
+     * variant.
+     *
+     * @param {Event} ev
+     * @param {Element} parent
+     * @param {Object} combination
      */
     _onChangeCombination(ev, parent, combination) {
         super._onChangeCombination(...arguments);
@@ -16,4 +19,4 @@ patch(WebsiteSale.prototype, {
             currencyValue.textContent = this._priceToStr(combination.l10n_ar_price_tax_excluded);
         }
     },
-})
+});
