+++ addons/website_sale_stock/__manifest__.py
@@ -30,7 +30,6 @@ Then it can be made specific at the product level.
     'auto_install': True,
     'assets': {
         'web.assets_frontend': [
-            ('before', 'website_sale/static/src/interactions/website_sale.js', 'website_sale_stock/static/src/js/variant_mixin.js'),
             'website_sale_stock/static/src/interactions/**/*',
             'website_sale_stock/static/src/js/combo_configurator_dialog/*',
             'website_sale_stock/static/src/js/models/*',
