+++ addons/website_sale_stock/static/src/interactions/website_sale.js
@@ -1,7 +1,11 @@
 import { rpc } from '@web/core/network/rpc';
 import { isEmail } from '@web/core/utils/strings';
 import { patch } from '@web/core/utils/patch';
+import { renderToFragment } from '@web/core/utils/render';
+import { formatFloat } from '@web/core/utils/numbers';
+import { setElementContent } from '@web/core/utils/html';
 import { patchDynamicContent } from '@web/public/utils';
+import { markup } from '@odoo/owl';
 import { WebsiteSale } from '@website_sale/interactions/website_sale';
 
 patch(WebsiteSale.prototype, {
@@ -64,12 +68,84 @@ patch(WebsiteSale.prototype, {
     },
 
     /**
-     * Adds the stock checking to the regular _onChangeCombination method
-     * @override
+     * Override of `website_sale` to check the product's stock.
+     *
+     * This will prevent the user from selecting a quantity that is not in stock for that product.
+     *
+     * It will also display various info/warning messages regarding the select product's stock.
+     *
+     * @param {Event} ev
+     * @param {Element} parent
+     * @param {Object} combination
      */
     _onChangeCombination(ev, parent, combination) {
         super._onChangeCombination(...arguments);
-        this._onChangeCombinationStock(...arguments);
+        const has_max_combo_quantity = 'max_combo_quantity' in combination
+        if (!combination.is_storable && !has_max_combo_quantity) return;
+        if (!parent.matches('.js_main_product') || !combination.product_id) {
+            // if we're not on product page or the product is dynamic
+            return;
+        }
+
+        const addQtyInput = parent.querySelector('input[name="add_qty"]');
+        let qty = addQtyInput.value;
+        const ctaWrapper = parent.querySelector('#o_wsale_cta_wrapper');
+        ctaWrapper.classList.replace('d-none', 'd-flex');
+        ctaWrapper.classList.remove('out_of_stock');
+
+        if (!combination.allow_out_of_stock_order) {
+            combination.free_qty -= parseInt(combination.cart_qty);
+            addQtyInput.dataset.max = combination.free_qty || 1;
+            if (combination.free_qty < 0) {
+                combination.free_qty = 0;
+            }
+            if (qty > combination.free_qty) {
+                qty = combination.free_qty || 1;
+                addQtyInput.value = qty;
+            }
+            if (combination.free_qty < 1) {
+                ctaWrapper.classList.replace('d-flex', 'd-none');
+                ctaWrapper.classList.add('out_of_stock');
+            }
+        }
+
+        if (has_max_combo_quantity) {
+            addQtyInput.dataset.max = combination.max_combo_quantity || 1;
+            if (qty > combination.max_combo_quantity) {
+                qty = combination.max_combo_quantity || 1;
+                addQtyInput.value = qty;
+            }
+            if (combination.max_combo_quantity < 1) {
+                ctaWrapper.classList.replace('d-flex', 'd-none');
+                ctaWrapper.classList.add('out_of_stock');
+            }
+        }
+
+        // needed xml-side for formatting of remaining qty
+        combination.formatQuantity = (qty) => {
+            if (Number.isInteger(qty)) {
+                return qty;
+            } else {
+                const decimals = Math.max(
+                    0,
+                    Math.ceil(-Math.log10(combination.uom_rounding))
+                );
+                return formatFloat(qty, {digits: [false, decimals]});
+            }
+        }
+
+        document.querySelector('.oe_website_sale')
+            .querySelectorAll('.availability_message_' + combination.product_template)
+            .forEach(el => el.remove());
+        if (combination.out_of_stock_message) {
+            combination.out_of_stock_message = markup(combination.out_of_stock_message);
+            const outOfStockMessage = document.createElement('div');
+            setElementContent(outOfStockMessage, combination.out_of_stock_message);
+            combination.has_out_of_stock_message = !!outOfStockMessage.textContent.trim();
+        }
+        this.el.querySelector('div.availability_messages').append(renderToFragment(
+            'website_sale_stock.product_availability', combination
+        ));
     },
 
     /**
