+++ addons/website_sale_stock_wishlist/static/src/interactions/website_sale.js
@@ -1,4 +1,5 @@
 import { patch } from '@web/core/utils/patch';
+import { renderToElement } from '@web/core/utils/render';
 import { patchDynamicContent } from '@web/public/utils';
 import { WebsiteSale } from '@website_sale/interactions/website_sale';
 
@@ -23,4 +24,27 @@ patch(WebsiteSale.prototype, {
         const productId = ev.currentTarget.closest('article').dataset.productId;
         this._handleClickSubmitStockNotificationForm(ev, productId);
     },
+
+    /**
+     * Override of `website_sale` to display additional info messages regarding the product's stock
+     * and the wishlist.
+     *
+     * @param {Event} ev
+     * @param {Element} parent
+     * @param {Object} combination
+     */
+    _onChangeCombination(ev, parent, combination) {
+        super._onChangeCombination(...arguments);
+        if (this.el.querySelector('.o_add_wishlist_dyn')) {
+            const messageEl = this.el.querySelector('div.availability_messages');
+            if (messageEl && !this.el.querySelector('#stock_wishlist_message')) {
+                this.services['public.interactions'].stopInteractions(messageEl);
+                messageEl.append(
+                    renderToElement('website_sale_stock_wishlist.product_availability', combination)
+                    || ''
+                );
+                this.services['public.interactions'].startInteractions(messageEl);
+            }
+        }
+    }
 });
